feat!: replace Engine with reactive counterpart and get rid of dependency graph
fix: remove obsolete reactive engine loading
chore(engine): use explicit synchronisation instead of thread contexts
