--- alchemist-engine/src/test/kotlin/it/unibo/alchemist/core/TestReactiveEngineImplementation.kt
@@ -1,12 +1,0 @@
-/*
- * Copyright (C) 2010-2026, Danilo Pianini and contributors
- * listed, for each module, in the respective subproject's build.gradle.kts file.
- *
- * This file is part of Alchemist, and is distributed under the terms of the
- * GNU General Public License, with a linking exception,
- * as described in the file LICENSE in the Alchemist distribution's top directory.
- */
-
-package it.unibo.alchemist.core
-
-class TestReactiveEngineImplementation : AbstractEngineTest({ env -> ReactiveEngine(env) })
