--- run_all_tests_reactive.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-./gradlew -Dalchemist.engine=reactive test
